backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
greenlet==3.3.2
playwright==1.58.0
psycopg2-binary==2.9.11
pyee==13.0.1
PyJWT==2.12.1
sqlparse==0.5.5
typing_extensions==4.15.0
gunicorn==23.0.0
django-environ==0.12.0
